+++ test/blackbox-tests/test-cases/variants/resolution-priority/bar.ml
@@ -1,0 +1,3 @@
+Direct.run ();;
+Variant.run ();;
+Test.run ();;