+++ test/blackbox-tests/test-cases/variants/resolution-priority/direct.default/direct.ml
@@ -1,0 +1,1 @@
+let run () = Printf.printf "hi from direct.default\n"