+++ test/blackbox-tests/test-cases/variants/variant-with-concrete-impl/lib.default/vlib.ml
@@ -1,0 +1,1 @@
+let run () = Printf.printf "hi from lib.default"